+++ Sources/StreamChatUI/CommonViews/AvatarView/ChatAvatarView.swift
@@ -5,7 +5,7 @@
 import UIKit
 
 /// A view that displays the avatar image. By default a circular image.
-open class ChatAvatarView: _View {
+open class ChatAvatarView: _Control {
     /// The `UIImageView` instance that shows the avatar image.
     open private(set) var imageView: UIImageView = UIImageView().withoutAutoresizingMaskConstraints